+++ AnalogClock/Definitions/UI/Theme/ClockElementColor.swift
@@ -1,0 +1,15 @@
+//
+//  ClockElementColor.swift
+//  AnalogClock
+//
+//  Created by Justin Reusch on 5/30/20.
+//  Copyright © 2020 Justin Reusch. All rights reserved.
+//
+
+import SwiftUI
+
+/// Holds color settings for fill and outline
+struct ClockElementColor {
+    var fill: Color? = nil
+    var outline: Color? = nil
+}
